+++ common/db/src/test/java/is/codion/common/db/operation/DefaultProcedureTypeTest.java
@@ -3,8 +3,6 @@
  */
 package is.codion.common.db.operation;
 
-import is.codion.common.db.exception.DatabaseException;
-
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -12,10 +10,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public final class DefaultProcedureTypeTest {
 
   @Test
-  void test() throws DatabaseException {
+  void test() {
     DefaultProcedureType<String, String> procedureType = new DefaultProcedureType<>("test");
     assertEquals(procedureType, new DefaultProcedureType<>("test"));
     assertEquals(procedureType.name(), procedureType.toString());
-    procedureType.execute("conn", (connection, argument) -> {}, "hello");
   }
 }